+++ src/main/java/com/iii/eeit124/member/service/MemberCenterService.java
@@ -1,5 +1,6 @@
 package com.iii.eeit124.member.service;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 import com.iii.eeit124.entity.Members;
@@ -7,5 +8,5 @@ import com.iii.eeit124.entity.Members;
 public interface MemberCenterService {
 	Members getMemberById(Integer memberId);
 
-	Map<String, Double> getDataPerMonth(Integer memberId);
+	Map<String, BigDecimal> getDataPerMonth(Integer memberId);
 }
